+++ lib/math_expressions.dart
@@ -33,9 +33,12 @@ library math_expressions;
 import 'dart:math' as math;
 
 import 'package:vector_math/vector_math.dart' show Vector2, Vector3, Vector4;
+import 'package:petitparser/petitparser.dart';
+import 'package:petitparser/petitparser.dart' as p;
 
 part 'src/algebra.dart';
 part 'src/expression.dart';
 part 'src/functions.dart';
 part 'src/parser.dart';
+part 'src/parser_petit.dart';
 part 'src/evaluator.dart';