+++ test/math_expressions_test.dart
@@ -11,6 +11,7 @@ import 'test_framework.dart';
 part 'algebra_test_set.dart';
 part 'expression_test_set.dart';
 part 'parser_test_set.dart';
+part 'parser_petit_test_set.dart';
 
 /// relative accuracy for floating-point calculations
 const num EPS = 0.000001;
@@ -20,6 +21,7 @@ void main() {
   final List<TestSet> testSets = <TestSet>[
     AlgebraTests(),
     ParserTests(),
+    PetitParserTests(),
     ExpressionTests()
   ];